# docs/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bank Cards Management",
    "contact": {
      "name": "Шаманова Елизавета Алексеевна",
      "email": "sandress2000@mail.ru"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/cards": {
      "get": {
        "tags": [
          "Card API"
        ],
        "summary": "Получить список карт пользователя",
        "description": "Возвращает информацию о картах с возможностью фильтрации по статусу и пагинации",
        "operationId": "getCardsInfo",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "UUID пользователя (опционально для администратора)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Статус карты для фильтрации",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "BLOCKED",
                "EXPIRED"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы пагинации",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Размер страницы пагинации",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 3
            },
            "example": 3
          }
        ],
        "responses": {
          "403": {
            "description": "Доступ запрещен"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список карт пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Card API"
        ],
        "summary": "Создать карту для пользователя",
        "description": "Создает карту, генерирует номер и шифрует его. Требуется ownerId и необязательно количество месяцев до истечения срока. Доступ есть только у админа",
        "operationId": "create",
        "parameters": [
          {
            "name": "ownerId",
            "in": "query",
            "description": "UUID владельца карты",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "monthsQuantityUntilExpires",
            "in": "query",
            "description": "Количество месяцев до окончания срока действия карты. Доступ есть только у админа.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Доступ запрещен"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Информация о созданной карте",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/cards/transfer": {
      "post": {
        "tags": [
          "Card API"
        ],
        "summary": "Перевод средств между картами",
        "description": "Выполняет перевод указанной суммы с одной карты на другую.",
        "operationId": "doMoneyTransfer",
        "requestBody": {
          "description": "Данные перевода",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardTransferMoney"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Карта недоступна для операции"
          },
          "400": {
            "description": "Некорректные данные запроса"
          },
          "404": {
            "description": "Одна из карт не найдена"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Перевод выполнен"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/cards/blocking/submit": {
      "post": {
        "tags": [
          "Card API"
        ],
        "summary": "Создать заявку на блокировку карты",
        "description": "Отправляет заявку на блокировку карты",
        "operationId": "submitCardBlocking",
        "requestBody": {
          "description": "Номер карты для блокировки",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardNumberBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Карта недоступна"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Заявка на блокировку создана"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/cards/blocking/process/{cardId}": {
      "post": {
        "tags": [
          "Card API"
        ],
        "summary": "Обработка заявки на блокировку карты",
        "description": "Меняет статус карты на BLOCKED и удаляет заявку на блокировку. Доступ есть только у админа.",
        "operationId": "processBlockRequest",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "UUID карты для блокировки",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Доступ запрещен"
          },
          "400": {
            "description": "Некорректный ID карты"
          },
          "404": {
            "description": "Карта не найдена"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Заявка обработана, карта заблокирована"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/cards/balance/view": {
      "post": {
        "tags": [
          "Card API"
        ],
        "summary": "Получить баланс карты",
        "description": "Возвращает текущий баланс карты по номеру",
        "operationId": "getBalance",
        "requestBody": {
          "description": "Номер карты",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardNumberBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Баланс карты",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardBalanceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/api/cards/{cardId}": {
      "delete": {
        "tags": [
          "Card API"
        ],
        "summary": "Удалить карту",
        "description": "Удаляет карту и связанные с ней данные. Доступ есть только у админа.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "UUID удаляемой карты",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Доступ запрещен"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Карта удалена"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Card API"
        ],
        "summary": "Обновить статус карты",
        "description": "Изменяет статус карты (ACTIVATE или BLOCK) по ID карты и действию. Доступ есть только у админа.",
        "operationId": "updateCardStatus",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "description": "Действие смены статуса карты",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVATE",
                "BLOCK"
              ]
            }
          },
          {
            "name": "cardId",
            "in": "path",
            "description": "UUID карты для обновления",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Доступ запрещен"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Статус карты обновлен"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/cards/balance/update": {
      "patch": {
        "tags": [
          "Card API"
        ],
        "summary": "Обновить баланс карты (пополнение/снятие)",
        "description": "Изменяет баланс карты в зависимости от действия WITHDRAW_MONEY или DEPOSIT_MONEY",
        "operationId": "updateCardBalance",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "description": "Действие обновления баланса",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DEPOSIT_MONEY",
                "WITHDRAW_MONEY"
              ]
            }
          }
        ],
        "requestBody": {
          "description": "Данные для обновления баланса",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Карта недоступна"
          },
          "400": {
            "description": "Некорректные данные или операция невозможна"
          },
          "404": {
            "description": "Карта не найдена"
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Баланс карты обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardBalanceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CardInfoResponse": {
        "type": "object",
        "description": "Ответ, содержащий информацию о карте",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID карты",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "maskedNumber": {
            "type": "string",
            "description": "Маскированный номер карты (для безопасности)",
            "example": 1234
          },
          "date": {
            "type": "string",
            "description": "Дата окончания срока действия карты в формате ММ/ГГ",
            "example": "12/26"
          },
          "status": {
            "type": "string",
            "description": "Статус карты",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED"
            ],
            "example": "ACTIVE"
          },
          "balance": {
            "type": "number",
            "format": "double",
            "description": "Баланс карты",
            "example": 2500.75
          },
          "owner": {
            "type": "string",
            "description": "Логин владельца карты",
            "example": "user123"
          }
        }
      },
      "CardTransferMoney": {
        "type": "object",
        "description": "Запрос на перевод денег между картами",
        "properties": {
          "from": {
            "type": "string",
            "description": "Номер карты отправителя",
            "example": "1234-5678-9012-3456"
          },
          "to": {
            "type": "string",
            "description": "Номер карты получателя",
            "example": "6543-2109-8765-4321"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "description": "Сумма перевода",
            "example": 150.0
          }
        },
        "required": [
          "amount",
          "from",
          "to"
        ]
      },
      "CardNumberBody": {
        "type": "object",
        "description": "Запрос, содержащий номер карты",
        "properties": {
          "number": {
            "type": "string",
            "description": "Номер карты",
            "example": 1234567890123456
          }
        }
      },
      "CardBalanceResponse": {
        "type": "object",
        "description": "Ответ с текущим балансом карты",
        "properties": {
          "balance": {
            "type": "number",
            "format": "double",
            "description": "Текущий баланс карты",
            "example": 2500.75
          }
        }
      },
...
    }
  }
}